Create 11-12semAlgorithms.cpp
There is a LOT of mistakes in there and I know that. I was coming up with better solution to the prev point but I didn't want to come back because I was really tired. I kept the programmer's oath: "If it works, don't fix it" -The Wise Unknown Programmer.